validators: extract the two existing validators; add test cases
Updates #cleanup
